+++ src/Segues/VerticalSlice.h
@@ -25,12 +25,12 @@ public:
     temp = new sf::Texture(surface.getTexture()); // Make a copy of the source texture
 
     sf::Sprite left(*temp); 
-    left.setTextureRect(sf::IntRect(0, 0, windowSize.x/2.0, windowSize.y));
-    left.setPosition(0, direction * alpha * left.getTexture()->getSize().y);
+    left.setTextureRect(sf::IntRect(0, 0, (int)(windowSize.x/2.0), windowSize.y));
+    left.setPosition(0, (float)(direction * alpha * (double)left.getTexture()->getSize().y));
 
     sf::Sprite right(*temp);
-    right.setTextureRect(sf::IntRect(windowSize.x/2.0, 0, windowSize.x, windowSize.y));
-    right.setPosition(windowSize.x/2.0, direction * -alpha * right.getTexture()->getSize().y);
+    right.setTextureRect(sf::IntRect((int)(windowSize.x/2.0), 0, windowSize.x, windowSize.y));
+    right.setPosition((float)(windowSize.x/2.0f), (float)(direction * -alpha * (double)right.getTexture()->getSize().y));
 
     surface.clear();
